+++ src/utils.ts
@@ -317,8 +317,13 @@ export const getResponseSchemaValidator = (
 	return record
 }
 
+const isBun = typeof Bun !== 'undefined'
+const hasHash = isBun && typeof Bun.hash === 'function'
+
 // https://stackoverflow.com/a/52171480
 export const checksum = (s: string) => {
+	if (hasHash) return Bun.hash(s) as number
+
 	let h = 9
 
 	for (let i = 0; i < s.length; ) h = Math.imul(h ^ s.charCodeAt(i++), 9 ** 9)